requirements.txt
anyio==4.4.0
asgiref==3.7.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
distlib==0.3.8
dj-database-url==2.3.0
Django==5.0.1
django-allauth==64.0.0
django-channels==0.7.0
django-cors-headers==4.6.0
django-filter==24.2
django-webpack-loader==3.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
mysql-connector-python==9.0.0
mysqlclient==2.2.1
npm==0.1.1
oauthlib==3.2.2
optional-django==0.1.0
packaging==24.2
pillow==10.3.0
pipenv==2023.12.0
platformdirs==4.2.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.8.0
python-telegram-bot==21.4
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==69.0.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.2
virtualenv==20.25.0
whitenoise==6.8.2

# Added from the upper list (if not already present)
amqp==5.3.1
billiard==4.2.1
celery==5.5.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
django-crontab==0.7.1
kombu==5.5.2
prompt_toolkit==3.0.50
python-dateutil==2.9.0.post0
redis==5.2.1
vine==5.1.0
wcwidth==0.2.13
joblib==1.5.1
pandas==2.3.1
scikit-learn==1.7.0